Change versioning scheme, new version is 51.5.0a0
In the new versioning scheme, the first and second components of the
version number are based on the underlying MonetDB version and the third
component is specific to monetdbe.

The previous version number was 0.11. In the tests we found references
to an even older versioning scheme with versions such as 3.8.8.
These have been removed.